# whatsapp_notification_open_wa_api_docs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/validate": {
      "post": {
        "operationId": "AuthValidateController_validate",
        "parameters": [
          {
            "name": "x-api-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API key to validate",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key is valid"
          },
          "401": {
            "description": "Invalid or missing API key"
          }
        },
        "summary": "Validate an API key",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenWA API",
    "description": "Open Source WhatsApp API Gateway - Free, Self-Hosted HTTP API",
    "version": "0.2.1",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
  },
  "security": [
    {
      "X-API-Key": []
    }
  ]
}